refactor(worktree): decompose GitPane god-component into hooks + panels + layout (#922)
Split the ~3.1k-line src/components/worktree/GitPane.tsx into a thin
coordinator that wires focused custom hooks (logic) to memo'd panel
components (presentation) via a shared layout — a behavior-preserving
refactor with the 112 existing GitPane tests as the safety net.

- Data/side-effects extracted to hooks: useGitStatus, useCommitHistory,
  useChanges, useBranches, useStash, useDangerZone (joining the existing
  useGitPaneNetworkOps / useGitPaneTabState). Cross-section refetch
  cascades are injected as callbacks, so the hooks stay decoupled and the
  coordinator composes them in topological order (no circular deps).
- Presentation extracted to git/panels/* (Status, Network, Changes,
  CommitHistory, Branch, Stash, DangerZone) + git/gitPaneShared, with the
  responsive mobile-tabs / desktop-grouping split moved to GitPaneLayout.
- Ambient config (isMobile / onDiffSelect / onInsertToMessage) supplied
  via GitPaneContext to eliminate prop-drilling into the panels.
- External props, data-testids, and runtime behavior are unchanged.
- Added focused unit tests for the new hooks and the context/layout.
